+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Skip updating asset_data.yml asset ID when running an update FROM the asset data
 * Updated defaults for CPU and RAM for template subscriptions, added a --ram arg to speficy the ram in a template subscription
+* Added the --url and --project args to the cons3rt CLI to allow selection of the URL/project API key when specified
 
 0.0.25
 ======